+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Builtin `!history`
 
+### Fixed
+- Path prefix change ending with `/` always resetting to the root (`/`)
+
 
 ## [0.7.0] - 2025-06-25
 ### Added